+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -8,7 +8,7 @@ CHANGELOG
    and `BusNameStamp` were added, which allow you to add a bus identifier
    to the `Envelope` then find the correct bus when receiving from
    the transport. See `ConsumeMessagesCommand`.
- * An optional `ConsumeMessagesCommand` constructor argument was removed.
+ * The optional `$busNames` constructor argument of the class `ConsumeMessagesCommand` was removed.
  * [BC BREAK] 2 new methods were added to `ReceiverInterface`:
    `ack()` and `reject()`.
  * [BC BREAK] Error handling was moved from the receivers into